backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.2
botocore==1.38.2
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
jmespath==1.0.1
mangum==0.19.0
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2

google-cloud-storage
google-cloud-firestore
